Tweaked the design a bit.
Now, internally in the room's main loop, we're maintaining references to the started and join channels.
When the limit is hit, the handle on the join channel is nilled out, blocking any more join requests
from being processed. The internal started handle reference is also assigned to a channel, until a message
has been sent down the channel and it is nilled out again.